+++ src/vmm/src/devices/virtio/net/metrics.rs
@@ -9,7 +9,7 @@
 //! ## JSON example with metrics:
 //! ```json
 //! {
-//!  "net": {
+//!  "net_eth0": {
 //!     "activate_fails": "SharedIncMetric",
 //!     "cfg_fails": "SharedIncMetric",
 //!     "mac_address_updates": "SharedIncMetric",
@@ -17,15 +17,7 @@
 //!     "no_tx_avail_buffer": "SharedIncMetric",
 //!     ...
 //!  }
-//!  "net0": {
-//!     "activate_fails": "SharedIncMetric",
-//!     "cfg_fails": "SharedIncMetric",
-//!     "mac_address_updates": "SharedIncMetric",
-//!     "no_rx_avail_buffer": "SharedIncMetric",
-//!     "no_tx_avail_buffer": "SharedIncMetric",
-//!     ...
-//!  }
-//!  "net1": {
+//!  "net_eth1": {
 //!     "activate_fails": "SharedIncMetric",
 //!     "cfg_fails": "SharedIncMetric",
 //!     "mac_address_updates": "SharedIncMetric",
@@ -34,7 +26,15 @@
 //!     ...
 //!  }
 //!  ...
-//!  "netN": {
+//!  "net_iface_id": {
+//!     "activate_fails": "SharedIncMetric",
+//!     "cfg_fails": "SharedIncMetric",
+//!     "mac_address_updates": "SharedIncMetric",
+//!     "no_rx_avail_buffer": "SharedIncMetric",
+//!     "no_tx_avail_buffer": "SharedIncMetric",
+//!     ...
+//!  }
+//!  "net": {
 //!     "activate_fails": "SharedIncMetric",
 //!     "cfg_fails": "SharedIncMetric",
 //!     "mac_address_updates": "SharedIncMetric",
@@ -46,7 +46,9 @@
 //! ```
 //! Each `net` field in the example above is a serializable `NetDeviceMetrics` structure
 //! collecting metrics such as `activate_fails`, `cfg_fails`, etc. for the network device.
-//! `net0`, `net1` and `netN` in the above example represent metrics 0th, 1st and 'N'th
+//! `net_eth0` represent metrics for the endpoint "/network-interfaces/eth0",
+//! `net_eth1` represent metrics for the endpoint "/network-interfaces/eth1", and
+//! `net_iface_id` represent metrics for the endpoint "/network-interfaces/{iface_id}"
 //! network device respectively and `net` is the aggregate of all the per device metrics.
 //!
 //! # Limitations
@@ -61,8 +63,13 @@
 //! * Use lockless operations, preferably ones that don't require anything other than simple
 //!   reads/writes being atomic.
 //! * Rely on `serde` to provide the actual serialization for writing the metrics.
-//! * Since all metrics start at 0, we implement the `Default` trait via derive for all of them, to
-//!   avoid having to initialize everything by hand.
+//!
+//! * Devices could be created in any order i.e. the first device created could either be eth0 or
+//!   eth1 so if we use a vector for NetDeviceMetrics and call 1st device as net0, then net0 could
+//!   sometimes point to eth0 and sometimes to eth1 which doesn't help with analysing the metrics.
+//!   So, use Map instead of Vec to help understand which interface the metrics actually belongs to.
+//! * We use "net_$iface_id" for the metrics name instead of "net_$tap_name" to be consistent with
+//!   the net endpoint "/network-interfaces/{iface_id}".
 //!
 //! The system implements 1 types of metrics:
 //! * Shared Incremental Metrics (SharedIncMetrics) - dedicated for the metrics which need a counter
@@ -70,80 +77,102 @@
 //! We use NET_DEV_METRICS_PVT instead of adding an entry of NetDeviceMetrics
 //! in Net so that metrics are accessible to be flushed even from signal handlers.
 
-use std::sync::RwLock;
+use std::collections::BTreeMap;
+use std::sync::{RwLock, RwLockReadGuard};
 
 use serde::ser::SerializeMap;
 use serde::{Serialize, Serializer};
 
 use crate::logger::{IncMetric, SharedIncMetric};
 
-#[derive(Copy, Clone, Debug, Default)]
-pub struct NetDeviceMetricsIndex(usize);
-
-impl NetDeviceMetricsIndex {
-    pub fn get(&self) -> &RwLock<NetDeviceMetrics> {
-        // SAFETY: This is called by individual Net devices
-        // on self (i.e. self.metrics.get()) so no race condition
-        // and it is safe.
-        unsafe { &NET_DEV_METRICS_PVT.metrics[self.0] }
-    }
+/// returns lock-unwrapped NET_DEV_METRICS_PVT.metrics,
+/// we have this function so that we don't have to make
+/// NET_DEV_METRICS_PVT public
+pub fn get_net_metrics() -> RwLockReadGuard<'static, NetMetricsPerDevice> {
+    // lock is always initialized so it is safe the unwrap the lock without a check.
+    NET_DEV_METRICS_PVT.read().unwrap()
 }
 
-/// provides instance for net metrics
+/// map of network interface id and metrics
+/// this should be protected by a lock before accessing.
 #[derive(Debug)]
-pub struct NetDeviceMetricsAllocator {
+pub struct NetMetricsPerDevice {
     /// used to access per net device metrics
-    metrics: Vec<RwLock<NetDeviceMetrics>>,
+    pub metrics: BTreeMap<String, NetDeviceMetrics>,
 }
 
-impl NetDeviceMetricsAllocator {
-    /// default construction
-    pub fn alloc() -> NetDeviceMetricsIndex {
-        // SAFETY: This is only called during creation of a Net device
-        // and since each device is created sequentially so there is no
-        // case of race condition.
-        unsafe {
+impl NetMetricsPerDevice {
+    /// Allocate `NetDeviceMetrics` for net device having
+    /// id `iface_id`. Also, allocate only if it doesn't
+    /// exist to avoid overwriting previously allocated data.
+    /// lock is always initialized so it is safe the unwrap
+    /// the lock without a check.
+    pub fn alloc(iface_id: String) {
+        if NET_DEV_METRICS_PVT
+            .read()
+            .unwrap()
+            .metrics
+            .get(&iface_id)
+            .is_none()
+        {
             NET_DEV_METRICS_PVT
+                .write()
+                .unwrap()
                 .metrics
-                .push(RwLock::new(NetDeviceMetrics::new()));
-            NetDeviceMetricsIndex(NET_DEV_METRICS_PVT.metrics.len() - 1)
+                .insert(iface_id, NetDeviceMetrics::new());
         }
     }
 }
 
-/// Contains Network-related metrics per device.
-static mut NET_DEV_METRICS_PVT: NetDeviceMetricsAllocator = NetDeviceMetricsAllocator {
-    metrics: Vec::new(),
-};
-
-pub fn flush_metrics<S: Serializer>(serializer: S) -> Result<S::Ok, S::Error> {
-    // SAFETY: This is called from vmm::logger::metrics::METRICS.write which
-    // has a lock and takes care of race condition.
-    unsafe {
-        // +1 to accomodate aggregate net metrics
-        let mut seq = serializer.serialize_map(Some(1 + NET_DEV_METRICS_PVT.metrics.len()))?;
-
-        let net_aggregated: NetDeviceMetrics = NET_DEV_METRICS_PVT.metrics.iter().fold(
-            NetDeviceMetrics::default(),
-            |mut net_agg, net| {
-                net_agg.aggregate(&net.read().unwrap());
-                net_agg
-            },
-        );
-
-        seq.serialize_entry("net", &net_aggregated)?;
-
-        for (i, metrics) in NET_DEV_METRICS_PVT.metrics.iter().enumerate() {
-            let devn = format!("net{}", i);
-            seq.serialize_entry(&devn, &*metrics.read().unwrap())?;
+#[macro_export]
+// Per device metrics are behind a ref lock and in a map which
+// makes it difficult to return the metrics pointer via a function
+// This macro is an attempt to make the code more readable and to
+// centralize the error handling.
+// The macro can be called in below ways where iface_id is String and
+// activate_fails is a metric from struct `NetDeviceMetrics`:
+//     NET_METRICS!(&iface_id, activate_fails.inc());
+//     NET_METRICS!(&iface_id, activate_fails.add(5));
+//     NET_METRICS!(&iface_id, activate_fails.count());
+macro_rules! NET_METRICS {
+    ($iface_id:expr,$metric:ident.$action:ident($($value:tt)?)) => {
+        if get_net_metrics().metrics.get($iface_id).is_some() {
+            get_net_metrics().metrics.get($iface_id).unwrap().$metric.$action($($value)?)
+        }else{
+            NetMetricsPerDevice::alloc($iface_id.to_string());
+            get_net_metrics().metrics.get($iface_id).unwrap().$metric.$action($($value)?)
         }
-        seq.end()
     }
+}
+
+/// Pool of Network-related metrics per device behing a lock to
+/// keep things thread safe. Since the lock is initialized here
+/// it is safe to unwrap it without any check.
+static NET_DEV_METRICS_PVT: RwLock<NetMetricsPerDevice> = RwLock::new(NetMetricsPerDevice {
+    metrics: BTreeMap::new(),
+});
+
+/// This function facilitates aggregation and serialization of
+/// per net device metrics.
+pub fn flush_metrics<S: Serializer>(serializer: S) -> Result<S::Ok, S::Error> {
+    let metrics_len = NET_DEV_METRICS_PVT.read().unwrap().metrics.len();
+    // +1 to accomodate aggregate net metrics
+    let mut seq = serializer.serialize_map(Some(1 + metrics_len))?;
+
+    let mut net_aggregated: NetDeviceMetrics = NetDeviceMetrics::new();
+
+    for metrics in NET_DEV_METRICS_PVT.read().unwrap().metrics.iter() {
+        let devn = format!("net_{}", metrics.0);
+        // serialization will flush the metrics so aggregate before it.
+        net_aggregated.aggregate(metrics.1);
+        seq.serialize_entry(&devn, &metrics.1)?;
+    }
+    seq.serialize_entry("net", &net_aggregated)?;
+    seq.end()
 }
 
 /// Network-related metrics.
-// #[derive(Debug, Default)]
-#[derive(Debug, Default, Serialize)]
+#[derive(Debug, Serialize)]
 pub struct NetDeviceMetrics {
     /// Number of times when activate failed on a network device.
     pub activate_fails: SharedIncMetric,
@@ -290,22 +319,61 @@ impl NetDeviceMetrics {
 #[cfg(test)]
 pub mod tests {
     use super::*;
+
     #[test]
-    fn test_net_dev_metrics() {
-        // we can have max 19 net devices
+    fn test_max_net_dev_metrics() {
+        // Note: this test has nothing to do with
+        // Net structure or IRQs, this is just to allocate
+        // metrics for max number of devices that system can have.
+        // we have 5-23 irq for net devices so max 19 net devices.
         const MAX_NET_DEVICES: usize = 19;
-        let mut net_dev_metrics: [NetDeviceMetricsIndex; MAX_NET_DEVICES] =
-            [NetDeviceMetricsIndex::default(); MAX_NET_DEVICES];
-        for metric in net_dev_metrics.iter_mut().take(MAX_NET_DEVICES) {
-            *metric = NetDeviceMetricsAllocator::alloc();
-            metric.get().write().unwrap().activate_fails.inc();
+
+        assert!(NET_DEV_METRICS_PVT.read().is_ok());
+        assert!(NET_DEV_METRICS_PVT.write().is_ok());
+
+        for i in 0..MAX_NET_DEVICES {
+            let devn: String = format!("eth{}", i);
+            NetMetricsPerDevice::alloc(devn.clone());
+            NET_METRICS!(&devn, activate_fails.inc());
+            NET_METRICS!(&devn, rx_bytes_count.add(10));
+            NET_METRICS!(&devn, tx_bytes_count.add(5));
         }
-        // SAFETY: something
-        unsafe {
-            assert!(NET_DEV_METRICS_PVT.metrics.len() >= MAX_NET_DEVICES);
+        for i in 0..MAX_NET_DEVICES {
+            let devn: String = format!("eth{}", i);
+            assert!(NET_METRICS!(&devn, activate_fails.count()) >= 1);
+            assert!(NET_METRICS!(&devn, rx_bytes_count.count()) >= 10);
+            assert_eq!(NET_METRICS!(&devn, tx_bytes_count.count()), 5);
         }
-        for metric in net_dev_metrics.iter_mut().take(MAX_NET_DEVICES) {
-            assert_eq!(metric.get().write().unwrap().activate_fails.count(), 1);
-        }
+    }
+    #[test]
+    fn test_signle_net_dev_metrics() {
+        // Use eth0 so that we can check thread safety with the
+        // `test_net_dev_metrics` which also uses the same name.
+        let devn = "eth0";
+
+        assert!(NET_DEV_METRICS_PVT.read().is_ok());
+        assert!(NET_DEV_METRICS_PVT.write().is_ok());
+
+        NetMetricsPerDevice::alloc(String::from(devn));
+        assert!(NET_DEV_METRICS_PVT.read().is_ok());
+        assert!(get_net_metrics().metrics.get(devn).is_some());
+
+        NET_METRICS!(devn, activate_fails.inc());
+        assert!(
+            NET_METRICS!(devn, activate_fails.count()) > 0,
+            "{}",
+            NET_METRICS!(devn, activate_fails.count())
+        );
+        // we expect only 2 tests (this and test_max_net_dev_metrics)
+        // to update activate_fails count for eth0.
+        assert!(
+            NET_METRICS!(devn, activate_fails.count()) <= 2,
+            "{}",
+            NET_METRICS!(devn, activate_fails.count())
+        );
+
+        NET_METRICS!(&String::from(devn), activate_fails.inc());
+        NET_METRICS!(&String::from(devn), rx_bytes_count.add(5));
+        assert!(NET_METRICS!(&String::from(devn), rx_bytes_count.count()) >= 5);
     }
 }