+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 
 ### Added
 
+- [#4145](https://github.com/firecracker-microvm/firecracker/pull/4145):
+  Added support for per net device metrics. Now, along with aggregated
+  network device metrics, metrics per network device with also be emitted.
+  Aggregate metrics will continue to be emitted with label "net" in the
+  metrics json object while each net device metrics will be emitted with
+  label "net_{iface_id}". E.g. the associated metrics for the endpoint
+  "/network-interfaces/eth0" will be available under "net_eth0:" in each
+  metrics json object.
+
 ### Changed
 
 - Simplified and clarified the removal policy of deprecated API elements
@@ -37,9 +46,6 @@
   capability checks that Firecracker performs during boot. If any of
   these fields are in use, minimal target snapshot version is
   restricted to 1.5.
-- [#4145](https://github.com/firecracker-microvm/firecracker/pull/4145):
-  Added support for per net device metrics. Now, along with aggregated
-  network device metrics, metrics per network device with also be emitted.
 
 ### Changed